better return value for stats
git-svn-id: https://svn-i2.informatik.rwth-aachen.de/repos/libalf/trunk@961 ee1db7db-6c55-0410-a809-e4d5348c71a5

+++ libalf/include/libalf/knowledgebase.h
@@ -436,11 +436,11 @@ class knowledgebase {
 					b = other->get_word();
 					return libalf::is_graded_lex_smaller(a,b);
 				}}}
-				int get_memory_usage()
+				unsigned long long int get_memory_usage()
 				// calculate memory consumption of this
 				// subtree
 				{{{
-					int ret;
+					unsigned long long int ret;
 					typename vector<node*>::iterator ci;
 
 					ret = sizeof(this) + sizeof(node *) * children.size();
@@ -913,9 +913,9 @@ class knowledgebase {
 			return true;
 		}}}
 
-		int get_memory_usage()
+		unsigned long long int get_memory_usage()
 		{{{
-			int ret;
+			unsigned long long int ret;
 
 			ret = sizeof(this);
 			ret += root->get_memory_usage();
